+++ pyvistaqt/plotting.py
@@ -78,6 +78,8 @@ from .utils import (
     _setup_application,
     _setup_ipython,
     _setup_off_screen,
+    pad_image,
+    resample_image,
 )
 from .window import MainWindow
 
@@ -104,31 +106,6 @@ LOG.addHandler(logging.StreamHandler())
 
 SAVE_CAM_BUTTON_TEXT = "Save Camera"
 CLEAR_CAMS_BUTTON_TEXT = "Clear Cameras"
-
-
-def resample_image(arr: np.ndarray, max_size: int = 400) -> np.ndarray:
-    """Resample a square image to an image of max_size."""
-    dim = np.max(arr.shape[0:2])
-    max_size = min(max_size, dim)
-    x_size, y_size, _ = arr.shape
-    s_x = int(np.ceil(x_size / max_size))
-    s_y = int(np.ceil(y_size / max_size))
-    img = np.zeros((max_size, max_size, arr.shape[2]), dtype=arr.dtype)
-    arr = arr[0:-1:s_x, 0:-1:s_y, :]
-    x_l = (max_size - arr.shape[0]) // 2
-    y_l = (max_size - arr.shape[1]) // 2
-    img[x_l : arr.shape[0] + x_l, y_l : arr.shape[1] + y_l, :] = arr
-    return img
-
-
-def pad_image(arr: np.ndarray, max_size: int = 400) -> np.ndarray:
-    """Pad an image to a square then resamples to max_size."""
-    dim = np.max(arr.shape)
-    img = np.zeros((dim, dim, arr.shape[2]), dtype=arr.dtype)
-    x_l = (dim - arr.shape[0]) // 2
-    y_l = (dim - arr.shape[1]) // 2
-    img[x_l : arr.shape[0] + x_l, y_l : arr.shape[1] + y_l, :] = arr
-    return resample_image(img, max_size=max_size)
 
 
 @contextlib.contextmanager
